indexdata: initial tombstone support
This commit adds repository tombstoning on a compound shard. The
implementation uses a boolean slice which has a 1-1 correspondance with
the repository list of a shard. Then at any point we make decisions (at
search time) we check if the repository for a document is alive.

This commit does not yet add any code which will actually tombstone a
repository. A follow-up commit can add that. The intention is the
tombstone array is immutable after loading to avoid synchronisation.

Change-Id: I581a1267271e1cae369a91610050b19496bb9e27