shards: support compound shards in getShards
Previously getShards would work as if only the first repository
existed (rankedShard only stored the first name). rankedShard now stores
a slice of names. We use the repository with the highest priority as the
priority for the shard.

Additionally the selectRepoSet optimization will work correctly for
compoundShards now. To make this work we had to exclude the optimization
where a compound shard contained repos not in the reposet/repobranch.

Change-Id: I010be6dc222f6960dd46b44b3d45f0aa97804295